updated websocket and kafka,redis classes

+++ quotes-service/src/main/java/com/market/quotesservice/dto/QuoteDto.java
@@ -12,5 +12,5 @@ public class QuoteDto {
     private double price;
     private double change;
     private double changePercent;
-    private long volume;
+    private double volume;
 }
